Sonarqubeの静的解析の警告修正

+++ Backend/src/Wellship.WebAPI/ExternalConnection/PostgreSQL/RepositoryImpls/StaffRepository.cs
@@ -43,7 +43,7 @@ namespace Ryobi.Wellship.WebAPI.ExternalConnection.PostgreSQL.RepositoryImpls
                         PasswordHash = s.PasswordHash,
                         PasswordSalt = s.PasswordSalt,
                         Enabled = s.Enabled,
-                        RoleId = (int)s.RoleId,
+                        RoleId = s.RoleId,
                         CreatedAt = createdAt,
                         CreatedBy = createdBy
                     }).ToArray();
